+++ crates/ov6_user_lib/src/thread.rs
@@ -1,5 +1,5 @@
 use crate::os::ov6::syscall;
 
-pub fn sleep(dur: i32) {
+pub fn sleep(dur: u64) {
     syscall::sleep(dur)
 }
